+++ drift/lib/drift3.dart
@@ -3,6 +3,16 @@ export 'src/dsl/database.dart';
 export 'src/dsl/table.dart';
 
 export 'src/query_builder/dialect.dart';
+
+export 'src/query_builder/schema/column.dart';
+export 'src/query_builder/schema/entities.dart';
+export 'src/query_builder/schema/result_set.dart';
+export 'src/query_builder/schema/table.dart';
+export 'src/query_builder/schema/view.dart';
+
+export 'src/query_builder/statements/select.dart';
+
+export 'src/query_builder/results.dart';
 export 'src/query_builder/types.dart';
 
 export 'src/runtime/streams/update_rules.dart';