+++ drift/lib/src/drift_dev_helper.dart
@@ -2,10 +2,13 @@
 export 'dart:typed_data' show Uint8List;
 
 export 'runtime/type_converter.dart' show TypeConverter, JsonTypeConverter2;
-export 'runtime/types/mapping.dart' show DriftAny, UserDefinedSqlType;
 export 'runtime/query_builder/query_builder.dart' show TableInfo;
 
-export 'dsl/table.dart' show Table, TableIndex, View;
-export 'dsl/database.dart' show DriftAccessor, DriftDatabase;
-
 export '../extensions/geopoly.dart';
+
+// new drift3
+
+export 'dsl/table.dart'
+    show DriftColumnDeclarationBuilder, Table, TableIndex, View;
+export 'dsl/database.dart' show DriftAccessor, DriftDatabase;
+export 'query_builder/types.dart' show SqlType;