+++ drift_dev/test/analysis/test_utils.dart
@@ -9,7 +9,7 @@ import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/file_system/overlay_file_system.dart';
 import 'package:analyzer/file_system/physical_file_system.dart';
 import 'package:build/build.dart';
-import 'package:drift/drift.dart';
+import 'package:drift/drift3.dart';
 import 'package:drift_dev/src/analysis/backend.dart';
 import 'package:drift_dev/src/analysis/driver/driver.dart';
 import 'package:drift_dev/src/analysis/driver/error.dart';
@@ -292,7 +292,7 @@ class TestImportManager implements ImportManager {
 Matcher get hasNoErrors =>
     isA<FileState>().having((e) => e.allErrors, 'allErrors', isEmpty);
 
-Matcher returnsColumns(Map<String, DriftSqlType> columns) {
+Matcher returnsColumns(Map<String, BuiltinDriftType> columns) {
   return _HasInferredColumnTypes(columns);
 }
 
@@ -309,7 +309,10 @@ class _HasInferredColumnTypes extends CustomMatcher {
     final resultSet = actual.resultSet;
     return {
       for (final column in resultSet.scalarColumns)
-        column.name: column.sqlType.builtin
+        column.name: switch (column.sqlType) {
+          ColumnDriftType(:final builtin) => builtin,
+          ColumnCustomType(:final custom) => 'custom: ${custom.expression}',
+        }
     };
   }
 }